+++ src/Storage/Storage/Blob/Cmdlet/SetAzureStorageBlobContent.cs
@@ -369,14 +369,14 @@ namespace Microsoft.WindowsAzure.Commands.Storage.Blob
                 Tuple<string, StorageBlob.CloudBlob> uploadRequest = UploadRequests.DequeueRequest();
                 IStorageBlobManagement localChannel = Channel;
                 Func<long, Task> taskGenerator;
-                if (!UseTrack2SDK())
+                if (!UseTrack2Sdk())
                 {
                     //Upload with DMlib
                     taskGenerator = (taskId) => Upload2Blob(taskId, localChannel, uploadRequest.Item1, uploadRequest.Item2);
                 }
                 else
                 {
-                   taskGenerator = (taskId) => UploadBlobwithSDK(taskId, localChannel, uploadRequest.Item1, uploadRequest.Item2);
+                   taskGenerator = (taskId) => UploadBlobwithSdk(taskId, localChannel, uploadRequest.Item1, uploadRequest.Item2);
                 }
                 RunTask(taskGenerator);
             }
@@ -414,7 +414,7 @@ namespace Microsoft.WindowsAzure.Commands.Storage.Blob
         /// <summary>
         /// Upload File to blob with storage Client library API
         /// </summary>
-        internal virtual async Task UploadBlobwithSDK(long taskId, IStorageBlobManagement localChannel, string filePath, StorageBlob.CloudBlob blob)
+        internal virtual async Task UploadBlobwithSdk(long taskId, IStorageBlobManagement localChannel, string filePath, StorageBlob.CloudBlob blob)
         {
             BlobClientOptions options = null;