+++ src/Storage/Storage.Management/ChangeLog.md
@@ -18,7 +18,14 @@
         - Additional information about change #1
 -->
 ## Upcoming Release
-
+* Supported create container/blob Sas token with new permission x,t
+    -  `New-AzStorageBlobSASToken`
+    -  `New-AzStorageContainerSASToken`
+* Supported create account Sas token with new permission x,t,f
+    -  `New-AzStorageAccountSASToken`
+* Supported get single file share usage
+    - `Get-AzRmStorageShare`
+        
 ## Version 2.3.0
 * Fixed the issue that UserAgent is not added for some data plane cmdlets.
 * Supported create/update Storage account with MinimumTlsVersion and AllowBlobPublicAccess